adds @builder, @getter, and @setter annotations
