+++ src/main/java/sai_adapa/projs/inv_management/auth/AdminVerifierInterceptor.java
@@ -1,14 +1,14 @@
 package sai_adapa.projs.inv_management.auth;
 
-import org.springframework.beans.factory.annotation.Configurable;
 import org.springframework.web.servlet.HandlerInterceptor;
 import sai_adapa.projs.inv_management.services.AdminService;
 import sai_adapa.projs.inv_management.tools.SpringContext;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
-@Configurable
+
 public class AdminVerifierInterceptor implements HandlerInterceptor {
 
 
@@ -19,7 +19,17 @@ public class AdminVerifierInterceptor implements HandlerInterceptor {
 
     @Override //Todo: add error resp in case of failure
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        return getAdminService().verifyToken(request.getHeader("session_token"));
+        if (getAdminService().verifyToken(request.getHeader("session_token"))) {
+            System.out.println("verified");
+            return true;
+        }
+        try {
+            response.getWriter().write("You are not an admin");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        System.out.println("not ver");
+        response.setStatus(401);
+        return false;
     }
-
 }
