+++ src/main/java/sai_adapa/projs/inv_management/auth/UsersVerifierInterceptor.java
@@ -6,6 +6,7 @@ import sai_adapa.projs.inv_management.tools.SpringContext;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
 public class UsersVerifierInterceptor implements HandlerInterceptor {
 
@@ -15,6 +16,16 @@ public class UsersVerifierInterceptor implements HandlerInterceptor {
 
     @Override //Todo: add error resp in case of failure
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        return getUsersService().verifySession(request.getHeader("session_token"));
+        if (getUsersService().verifySession(request.getHeader("session_token")))
+        {
+            return true;
+        }
+        try {
+            response.getWriter().write("You are not an normal user");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        response.setStatus(401);
+        return false;
     }
 }
