+++ src/main/java/sai_adapa/projs/inv_management/auth/VendorVerifierInterceptor.java
@@ -6,6 +6,7 @@ import sai_adapa.projs.inv_management.tools.SpringContext;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
 public class VendorVerifierInterceptor implements HandlerInterceptor {
     private VendorService getVendorService() {
@@ -14,6 +15,14 @@ public class VendorVerifierInterceptor implements HandlerInterceptor {
 
     @Override ///Todo: add error resp in case of failure
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        return getVendorService().verifySession(request.getHeader("session_token"));
+        if(getVendorService().verifySession(request.getHeader("session_token")))
+            return true;
+        try {
+            response.getWriter().write("You are not a vendor");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        response.setStatus(401);
+        return false;
     }
 }
