+++ src/main/java/sai_adapa/projs/inv_management/controllers/ItemController.java
@@ -1,0 +1,4 @@
+package sai_adapa.projs.inv_management.controllers;
+
+public class ItemController {
+}
