+++ src/main/java/sai_adapa/projs/inv_management/controllers/StockController.java
@@ -1,0 +1,5 @@
+package sai_adapa.projs.inv_management.controllers;
+
+public class StockController {
+
+}
