+++ src/main/java/sai_adapa/projs/inv_management/controllers/UsersController.java
@@ -1,7 +1,7 @@
 package sai_adapa.projs.inv_management.controllers;
 
 import org.springframework.web.bind.annotation.*;
-import sai_adapa.projs.inv_management.users.io.PreUsers;
+import sai_adapa.projs.inv_management.entities.users.io.PreUsers;
 import sai_adapa.projs.inv_management.services.UsersService;
 
 @RestController
@@ -13,7 +13,7 @@ public class UsersController {
         this.usersService = usersService;
     }
 
-    @RequestMapping(method = RequestMethod.POST, value = {"/app_user"})
+    @RequestMapping(method = RequestMethod.POST, value = {"/app_user/new"})
     public void signUp(@RequestBody PreUsers preUser) {
         usersService.addUser(preUser.getName(), preUser.getEmail(), preUser.getDetails(), preUser.getPasswd());
     }
@@ -27,8 +27,8 @@ public class UsersController {
     }
 
     @RequestMapping(method = RequestMethod.PATCH, value = {"/app_user"})
-    public void editUser(@RequestBody PreUsers preUsers, @RequestHeader("session_token") String key) {
-        usersService.editUser(usersService.getUsersBySession(key), preUsers.getName(), preUsers.getEmail(), preUsers.getDetails(), preUsers.getPasswd());
+    public void editUser(@RequestBody PreUsers preUsers) {
+        usersService.editUser(usersService.getUser(preUsers.getEmail()), preUsers.getName(), preUsers.getEmail(), preUsers.getDetails(), preUsers.getPasswd());
     }
