+++ src/main/java/sai_adapa/projs/inv_management/entities/users/Users.java
@@ -1,0 +1,37 @@
+package sai_adapa.projs.inv_management.entities.users;
+
+import lombok.*;
+import org.hibernate.annotations.GenericGenerator;
+import org.hibernate.annotations.Type;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Users {
+
+    @Id
+    @GeneratedValue(generator = "uuid")
+    @GenericGenerator(name = "uuid", strategy = "uuid2")
+    @Column(unique = true, nullable = false)
+    @Type(type = "pg-uuid")
+    private Long user_id;
+
+    private String name;
+    @Column(unique = true)
+    private String email;
+    private String details;
+    private String passwdHash;
+    private String sessionToken;
+
+}
+
+
+
