+++ src/main/java/sai_adapa/projs/inv_management/entities/users/Vendor.java
@@ -1,9 +1,7 @@
-package sai_adapa.projs.inv_management.users;
+package sai_adapa.projs.inv_management.entities.users;
 
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+
+import lombok.*;
 import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.Type;
 
@@ -11,30 +9,29 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import java.util.UUID;
 
 
-@Builder
 @Entity
 @Getter
+@Setter
+@Builder
 @NoArgsConstructor
-public class Admin {
-
+@AllArgsConstructor
+public class Vendor {
 
     @Id
     @GeneratedValue(generator = "uuid")
     @GenericGenerator(name = "uuid", strategy = "uuid2")
     @Column(unique = true, nullable = false)
     @Type(type = "pg-uuid")
-    private UUID admin_id;
-    @Column(unique = true)
-    @Setter
-    private String email;
-    @Setter
-    private String passwdHash;
-    @Setter
-    @Column(unique = true, nullable = true)
-    private String sessionToken;
+    private Long vendor_id;
 
+    private String name;
+
+    @Column(unique = true)
+    private String email;
+    private String description;
+    private String passwdHash;
+    private String sessionToken;
 
 }
