+++ src/main/java/sai_adapa/projs/inv_management/entities/users/io/PreAdmin.java
@@ -1,0 +1,19 @@
+package sai_adapa.projs.inv_management.entities.users.io;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class PreAdmin {
+    private String email;
+    private String passwd;
+
+
+
+
+}
