+++ src/main/java/sai_adapa/projs/inv_management/entities/users/io/PreVendor.java
@@ -1,0 +1,20 @@
+package sai_adapa.projs.inv_management.entities.users.io;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class PreVendor {
+
+    private String name;
+    private String email;
+    private String description;
+    private String passwd;
+
+
+}
