--- src/main/java/sai_adapa/projs/inv_management/item/Item.java
@@ -1,48 +1,0 @@
-package sai_adapa.projs.inv_management.item;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class Item {
-
-    @Id
-    @GeneratedValue( generator = "identity")
-    private Long item_id;
-    private String name;
-    private String description;
-
-    public Item() {
-    }
-
-    public Item(String name, String description) {
-        this.name = name;
-        this.description = description;
-    }
-
-    public Long getItem_id() {
-        return item_id;
-    }
-
-    public void setItem_id(Long id) {
-        this.item_id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-}
