--- src/main/java/sai_adapa/projs/inv_management/item/ItemService.java
@@ -1,4 +1,0 @@
-package sai_adapa.projs.inv_management.item;
-
-public class ItemService {
-}
