--- src/main/java/sai_adapa/projs/inv_management/item/stock/StockController.java
@@ -1,5 +1,0 @@
-package sai_adapa.projs.inv_management.item.stock;
-
-public class StockController {
-
-}
