--- src/main/java/sai_adapa/projs/inv_management/item/stock/StockService.java
@@ -1,4 +1,0 @@
-package sai_adapa.projs.inv_management.item.stock;
-
-public class StockService {
-}
