+++ src/main/java/sai_adapa/projs/inv_management/repositories/ItemRepository.java
@@ -1,7 +1,8 @@
-package sai_adapa.projs.inv_management.item;
+package sai_adapa.projs.inv_management.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import sai_adapa.projs.inv_management.entities.items.Item;
 
 @Repository
 public interface ItemRepository extends JpaRepository<Item,String> {
