+++ src/main/java/sai_adapa/projs/inv_management/repositories/OrderRepository.java
@@ -1,0 +1,4 @@
+package sai_adapa.projs.inv_management.repositories;
+
+public class OrderRepository {
+}
