+++ src/main/java/sai_adapa/projs/inv_management/repositories/StockRepository.java
@@ -1,7 +1,8 @@
-package sai_adapa.projs.inv_management.item.stock;
+package sai_adapa.projs.inv_management.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import sai_adapa.projs.inv_management.entities.items.stock.Stock;
 
 @Repository
 public interface StockRepository extends JpaRepository<Stock,String> {
