+++ src/main/java/sai_adapa/projs/inv_management/services/ItemService.java
@@ -1,0 +1,4 @@
+package sai_adapa.projs.inv_management.services;
+
+public class ItemService {
+}
