--- src/main/java/sai_adapa/projs/inv_management/users/Users.java
@@ -1,83 +1,0 @@
-package sai_adapa.projs.inv_management.users;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-public class Users {
-
-    @Id
-    @GeneratedValue(generator = "identity")
-    private Long user_id;
-
-    private String name;
-    @Column(unique = true)
-    private String email;
-    private String details;
-    private String passwdHash;
-    private String sessionToken;
-
-    //constructors getters and setters
-    public Users() {
-    }
-
-    public Users(String name, String email, String details, String passwdHash) {
-        this.name = name;
-        this.email = email;
-        this.details = details;
-        this.passwdHash = passwdHash;
-    }
-
-    public String getSessionToken() {
-        return sessionToken;
-    }
-
-    public void setSessionToken(String sessionToken) {
-        this.sessionToken = sessionToken;
-    }
-
-    public String getPasswdHash() {
-        return passwdHash;
-    }
-
-    public void setPasswdHash(String passwdHash) {
-        this.passwdHash = passwdHash;
-    }
-
-    public Long getUser_id() {
-        return user_id;
-    }
-
-    public void setUser_id(Long user_id) {
-        this.user_id = user_id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String e_mail) {
-        this.email = e_mail;
-    }
-
-    public String getDetails() {
-        return details;
-    }
-
-    public void setDetails(String details) {
-        this.details = details;
-    }
-}
-
-
-
