--- src/main/java/sai_adapa/projs/inv_management/users/io/PreVendor.java
@@ -1,41 +1,0 @@
-package sai_adapa.projs.inv_management.users.io;
-
-public class PreVendor {
-
-    private String name;
-    private String email;
-    private String description;
-    private String passwd;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getPasswd() {
-        return passwd;
-    }
-
-    public void setPasswd(String passwd) {
-        this.passwd = passwd;
-    }
-}
